Fixed the crash when loading the plugin on Ardour binary versions from ardour.org
The problem was related with the two fader widgets used to control the in/out gain. These faders are drawed using a png image and this image is loaded using the function Cairo::ImageSurface::create_from_png() which is only avalaible if Cairo have PNG support enabled using the CAIRO_HAS_PNG_FUNCTIONS macro.
For some reason, ardour from ardour.org doesn't like using such function from Cairo and crashes when trying to execute it with an unhandled exception.
Now, the PNG image is drawed using Gdk::Pixbuf and coping it to a Cairo surface. This works fine for mine, tested with both variants of ardour (binary from ardour.org and ArchLinux package).


git-svn-id: https://svn.code.sf.net/p/eq10q/code/trunk@79 8f919145-41b5-4f04-b112-5f015527c22b